Alot of changes

+++ Lets-Run-client/letsRun/src/app/event/comment/comment.component.ts
@@ -1,8 +1,9 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { CommentModule } from 'src/app/models/comment.model';
 import { ActivatedRoute, ParamMap } from '@angular/router';
 import { CommentService } from '../comment.service';
+import { AuthService } from 'src/app/auth/auth.service';
 
 @Component({
   selector: 'app-comment',
@@ -14,16 +15,22 @@ export class CommentComponent implements OnInit {
   showEditCommentForm = false;
   editCommentForm: FormGroup;
 
+  userId;
+
+  @Output() deletedCommentEmiter = new EventEmitter<boolean>();
+  commentDeleted = false;
   @Input() comment: CommentModule;
   @Input() eventId: string;
 
 
-  constructor(private activeRoute: ActivatedRoute, private commentService: CommentService) { }
+  constructor(private activeRoute: ActivatedRoute, private commentService: CommentService, private authService: AuthService) { }
 
   ngOnInit() {
+    this.userId = this.authService.getUserId();
     this.editCommentForm = new FormGroup({
       'content': new FormControl(null, {validators: [Validators.required]}),
     });
+    
   }
 
   edit_Comment() {
@@ -32,7 +39,8 @@ export class CommentComponent implements OnInit {
     }
 
     this.commentService.editComment(this.eventId, this.editCommentForm.value.content, this.comment.id);
-
+    this.editCommentForm.reset();
+    this.commentService.getCommentsList(this.eventId);
   }
 
   showEditCommentInputArea() {
@@ -40,6 +48,11 @@ export class CommentComponent implements OnInit {
   }
 
   deleteComment() {
+    if(this.comment.author === this.userId) {
+      return;
+    }
     this.commentService.deleteComent(this.eventId, this.comment.id);
+    this.editCommentForm.reset();
+    this.commentDeleted = true;
   }
 }
